93 move searchfield under textfield
