+++ src/Components/TextField/Searchfield/Searchfield.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { View, TextInput } from 'react-native';
-import withTheme from '../../Theme/withTheme';
-import IconButton from '../IconButton/IconButton';
+import withTheme from '../../../Theme/withTheme';
+import IconButton from '../../IconButton/IconButton';
 import styles from './Searchfield.styles';
 
 class Searchfield extends Component {
