+++ src/Components/TextField/Searchfield/Searchfield.stories.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import { View } from 'react-native';
 
-import { Searchfield } from '../../';
+import { Searchfield } from '../../../';
 import { State, Store } from '@sambego/storybook-state';
-import Header from '../../storybook/components/Header';
-import Container from '../../storybook/components/Container';
-import { storiesOf } from '../../storybook/helpers/storiesOf';
+import Header from '../../../storybook/components/Header';
+import Container from '../../../storybook/components/Container';
+import { storiesOf } from '../../../storybook/helpers/storiesOf';
 
 const store = new Store({
   searchOne: '',
 });
 
-export default storiesOf('Components|Searchfield', module)
+export default storiesOf('Components|TextField/SearchField', module)
   .addParameters({ jest: ['Searchfield'] })
   .add('Simple', () => (
     <Container>
